% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schemas.R
\name{lacks_mandatory}
\alias{lacks_mandatory}
\alias{violates_schema}
\title{Test for missing mandatory columns/entries in a table according to a schema}
\usage{
lacks_mandatory(df, schema)

violates_schema(df, schema)
}
\arguments{
\item{df}{\code{data.frame} or \code{list}}

\item{schema}{\code{data.table} schema with fields set according to the google doc}
}
\value{
\code{character} with names of missing fields, or of length 0 if no missing
}
\description{
Test for missing mandatory columns/entries in a table according to a schema
}
\section{Functions}{
\itemize{
\item \code{violates_schema}: test for incorrectly-specified columns/entries. Returns a named \code{character} with hints about incorrect entries, if any.
}}

