% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PGEXContainer-class}
\alias{PGEXContainer-class}
\title{An S4 Class to hold gene expression with appropriately labelled cell/feature variables
Extends a \link{\code{SummarizedExperiment}}.}
\description{
An S4 Class to hold gene expression with appropriately labelled cell/feature variables
Extends a \link{\code{SummarizedExperiment}}.
}
\section{Slots}{

\describe{
\item{\code{pgex_version}}{\code{character} giving version string of the package that created the object}
}}

