% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEXContainer.R
\name{PGEXContainerProto}
\alias{PGEXContainerProto}
\alias{PGEX_promote}
\title{Construct a PGEX Container Prototype}
\usage{
PGEXContainerProto(summarizedExperiment = NULL, pgex_experiment = NULL,
  pgex_sample = NULL, geo_soft = NULL, ae_idf = NULL, ae_sdrf = NULL,
  fdata = NULL, exprs = NULL)

PGEX_promote(obj)
}
\arguments{
\item{summarizedExperiment}{\code{SummarizedExperiment}, currently mandatory}

\item{pgex_experiment}{mandatory experiment data}

\item{pgex_sample}{optional experiment data to populate \code{colData(summarizedExperiment)}}

\item{geo_soft}{optional GEO data}

\item{ae_idf}{optional ArrayExpress}

\item{ae_sdrf}{optional ArrayExpress}

\item{fdata}{feature data (ignored)}

\item{exprs}{expression data (ignored)}

\item{obj}{\code{PGEXContainerProto}}
}
\value{
object of class \code{PGEXContainerProto}
}
\description{
A \code{PGEXContainerProto} is a \code{PGEXContainer} that has not had its fields validated for consistency and validity.
This allows step-wise construction.
}
\section{Functions}{
\itemize{
\item \code{PGEX_promote}: Promote an object to a full-fledged (validity checked) PGEXContainer
}}

\examples{
## library(GEOquery)
## geoexp <- getGEO('GSE42268')
## se <- makeSummarizedExperimentFromExpressionSet(geoexp[[1]])
library(SummarizedExperiment)
se <- readRDS(system.file("tests", "testthat", "GSE42268_SummarizedExperiment.rds",
                          package='PenultimateGEXContainer'))
dats <- split_invariant(colData(se))
colData(se)<- dats$variant
## No batch info available
## Note that we need to be careful to set the right data types.
## No coercion is attempted
batch  <- rep(NA_character_, ncol(se))
## NA by default is logical, which violates schema

## Would be a factor by default
biosample <- as.character(colData(se)$characteristics_ch1)
rna_pg <-  colData(se)$characteristics_ch1.2
ncells_maybe  <- colData(se)$characteristics_ch1.1
ncells <- rep(NA_real_, length(ncells_maybe))
ncells[grepl('single cell', ncells_maybe)] <- 1
ncells[grepl('50 cells', ncells_maybe) ] <- 50
ncells[grepl('12 cells', ncells_maybe)] <- 12
ncells[grepl('10 pg', rna_pg) & grepl('total RNA', ncells_maybe)] <- 1.2
ncells[grepl('1 ug', rna_pg) & grepl('total RNA', ncells_maybe)] <- 120
ncells[grepl('averaged single cell lysis', rna_pg) & grepl('LIMprep', ncells_maybe)] <- 1
cell_cycle_maybe <- colData(se)$characteristics_ch1.3
pgex_sample <- DataFrame(pgex_batch=batch, pgex_biosample=biosample, pgex_ncells=ncells, pgex_filter_study_pass=TRUE)
pgex_experiment <- list(pgex_platform='cell_ina_well', pgex_chemistry='3prime polyA tailing',
                        pgex_has_umi=FALSE, pgex_ercc_version=NA, pgex_molecule='polyA RNA',
                        pgex_is_public=TRUE, pgex_transformation=NA)
proto <- PGEXContainerProto(se, geo_soft=colData(se), pgex_sample=pgex_sample, pgex_experiment=pgex_experiment)
PGEX_promote(proto)
}
