% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEXContainer.R
\name{PGEXContainerProto}
\alias{PGEXContainerProto}
\title{Construct a PGEX Container Prototype}
\usage{
PGEXContainerProto(summarizedExperiment = NULL, pgex_experiment = NULL,
  pgex_sample = NULL, geo_soft = NULL, ae_idf = NULL, ae_sdrf = NULL,
  fdata = NULL, exprs = NULL)
}
\value{
object of class \code{PGEXContainerProto}
}
\description{
A \code{PGEXContainerProto} is a \code{PGEXContainer} that has not had its fields validated for consistency and validity.
This allows step-wise construction.
}
\examples{
\dontrun{
library(GEOquery)
geoexp <- getGEO('GSE42268')
se <- makeSummarizedExperimentFromExpressionSet(geoexp[[1]])
}
library(SummarizedExperiment)
se <- readRDS(system.file("tests", "testthat", "GSE42268_SummarizedExperiment.rds",
package='PenultimateGEXContainer'))
batch  <- rep(1, ncol(se))
biosample <- colData(se)$characteristics_ch1
rna_pg <-  colData(se)$characteristics_ch1.2
ncells_maybe  <- colData(se)$characteristics_ch1.1
ncells <- rep(NA_real_, length(ncells_maybe))
ncells[ncells_maybe \%like\% 'single cell'] <- 1
ncells[ncells_maybe \%like\% '50 cells'] <- 50
ncells[ncells_maybe \%like\% '12 cells'] <- 12
cell_cycle_maybe <- colData(se)$characteristics_ch1.3
cell_cycle <- colData(se)$characteristics_ch1.3
pgex_sample <- DataFrame(batch=rep(1, ncol(se)), biosample=rep(1, ncol(se)), hoechst_cell_cycle)
cont <- PGEXContainerProto(se, geoSoft=colData(se))
}
