% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEXContainer.R
\name{PGEXContainerPrototype}
\alias{PGEXContainerPrototype}
\title{Construct a PGEX Container Prototype}
\usage{
PGEXContainerPrototype(summarizedExperiment, pgex_experiment = DataFrame(),
  pgex_sample, geo_soft = NULL, ae_idf = NULL, ae_sdrf = NULL, cdata,
  fdata, exprs)
}
\value{
object of class \code{PGEXContainerPrototype}
}
\description{
A \code{PGEXContainerPrototype} is a \code{PGEXContainer} that has not had its fields validated for consistency and validity.
This allows step-wise construction.
}
\examples{
library(GEOquery)
geoexp <- getGEO('GSE42268')
se <- lapply(geoexp, makeSummarizedExperimentFromExpressionSet)
geoDF <- pData(geoexp[[1]])
batch  <- rep(1, ncol(se[[1]]))
biosample <- colData(se[[1]])$characteristics_ch1
rna_pg <-  colData(se[[1]])$characteristics_ch1.2
ncells_maybe  <- colData(se[[1]])$characteristics_ch1.1
ncells <- rep(NA_numeric_, length(ncells_maybe))
ncells[ncells_maybe \%like\% 'single cell'] <- 1
ncells[ncells_maybe \%like\% '50 cells'] <- 50
ncells[ncells_maybe \%like\% '12 cells'] <- 12
cell_cycle_maybe <- colData(se[[1]])$characteristics_ch1.3
cell_cycle <- colData(se[[1]])$characteristics_ch1.3
pgex_sample <- DataFrame(batch=rep(1, ncol(se[[1]])), biosample=rep(1, ncol(se[[1]])), hoechst_cell_cycle
cont <- PGEXContainerPrototype(se[[1]], geoSoft=colData(se[[1]]))
}
