% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGEXContainer.R
\name{split_invariant}
\alias{split_invariant}
\title{Separate variant and invariant columns}
\usage{
split_invariant(df)
}
\arguments{
\item{df}{\code{data.frame}}
}
\value{
\code{list} with entries \code{data.frame} \code{variant} and named list \code{constants}
}
\description{
Pull off columns from a \code{data.frame} that are constant
}
